Change Config class to StorageFactory
