+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 2.0.0
-======
+=====
+
+  * Change `Config` class to `StorageFactory`
+  * Remove the `get`, `set` and `__construct` methods in the `StorageFactory` class
+  * Move the `StorageFactory` class to the `SolidWorx\Toggler\Storage` namespace
+
+2.0.0 Alpha 1
+=============
 
   * Add a Symfony command to update the status of a feature
   * Added Symfony command to get the status of a feature
